rename viz module
