+++ moabb/tests/__init__.py
@@ -1,4 +1,4 @@
-from moabb.tests.viz import *
+from moabb.tests.analysis import *
 from moabb.tests.datasets import *
 from moabb.tests.contexts import *
 from moabb.tests.util_tests import *
